requirements.txt
pynacl==1.5.0
discord.py[voice]==2.2.2
python-dotenv==1.0.0